+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Display a welcome page on first extension installation.
 
 ### Changed
-- Automatically refresh YouTube and YouTube No-Cookie tabs upon first extension installation.
+- Automatically refresh open YouTube and YouTube No-Cookie tabs upon first extension installation.
+- Display the extension version dynamically in the popup and welcome page.
 
 ## [2.3.12] - 2025-06-14